pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copom-streamlit"
version = "0.1.0"
description = "Streamlit UI for the COPOM RAG system"
requires-python = ">=3.11"

dependencies = [
    "streamlit>=1.35.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]
